Fix stack display problem where stack is at a page boundary.
When mapping application space into Z80 space need to map the SP address, not the address after adjusting upwards. P_MAPX maps the page containing the address AND THE NEXT page into banks 1 and 2. If the stack is on a boundary and then we add 15 and convert to Z80 space it maps the wrong page.